Plugboard implementation in header class
Implemented the plugboard in the VEM wich beheaves in the same way as the mechanic one. Solved a bug in the Rotor class wich didn't created a new line if the rotor didn't recongize the character when encoding it. In addition the default rotors order was changed to a more realistic one.